+++ common/utils/utils.go
@@ -506,12 +506,3 @@ func DialectToTarget(dialect string) string {
 	}
 	return constants.TargetSpanner
 }
-
-func Contains(l []string, str string) bool {
-	for _, s := range l {
-		if s == str {
-			return true
-		}
-	}
-	return false
-}
