--- webv2/utility/utility.go
@@ -1,64 +1,0 @@
-// Copyright 2022 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-// Package web defines web APIs to be used with harbourbridge frontend.
-// Apart from schema conversion, this package involves API to update
-// converted schema.
-
-package utility
-
-// ContainString check string is present in given list.
-func ContainString(fc []string, col string) string {
-
-	for _, s := range fc {
-		if s == col {
-			return col
-		}
-	}
-	return ""
-}
-
-// DuplicateInArray checks if there is any duplicate element present in the list.
-func DuplicateInArray(element []int) int {
-	visited := make(map[int]bool, 0)
-	for i := 0; i < len(element); i++ {
-		if visited[element[i]] == true {
-			return element[i]
-		} else {
-			visited[element[i]] = true
-		}
-	}
-	return -1
-}
-
-// Difference gives list of element that are only present in first list.
-func Difference(listone, listtwo []int) []int {
-
-	hashmap := make(map[int]int, len(listtwo))
-
-	for _, val := range listtwo {
-		hashmap[val]++
-	}
-
-	var diff []int
-
-	for _, val := range listone {
-
-		_, found := hashmap[val]
-		if !found {
-			diff = append(diff, val)
-		}
-	}
-	return diff
-}
